feat: support fastify v4
